Support specifying the output format of fetch features with a query parameter
Add a servlet Filter that replaces the `Accept` request header by the MimeType of
a matching collection items `MimeTypes` supported format, if the
`f` query parameter matches a `MimeTypes#getShortName()` short
name supported output format.

E.g.: `GET /ogcapi/collections/{collectionId}/items?f=shapefile`
matches `MimeTypes#SHAPEFILE` and the Accept header is set to
`application/x-shapefile`, resulting in a download of the collection
in shapefile format and not in an error.