Use OpenAPI tag names to generage Java API interfaces
Add the `<useTags>true</useTags>` configuration option to the
`openapi-generator-maven-plugin`, so it generates Java API interfaces
based on the tags of each operation instead of using the URI paths.

This leads to the following refactoring of classes:

`CollectionsApi` -> `DataApi`
`DefaultApi` -> `CapabilitiesApi`

Additionally, some methods from the `DataApi` are moved to the
`CapabilitiesApi`, as they're annotated with the `Capabilities` tag in
the OpenAPI definition.